aligned official results with eval-arqmath1-task2.

+++ eval-arqmath1-task2/eval.sh
@@ -1,5 +1,5 @@
 QREL="topics-and-qrels/qrels.arqmath-2020-task2-visual_ids.v3.txt"
-TSV="./latex_representation_v2"
+TSV="./latex_representation_v3"
 DIR=$(dirname $0)
 
 for arg in "$@"; do
@@ -13,7 +13,7 @@ for arg in "$@"; do
       shift
       ;;
     --tsv=*)
-      TSV="${arg#*=}/"
+      TSV="${arg#*=}"
       shift
       ;;
     -*|--*)
@@ -26,22 +26,29 @@ for arg in "$@"; do
 done
 
 mkdir -p $DIR/prime-output
-rm -f $DIR/prime-output/*
-
 wc -l $DIR/input/*
 
 set -ex
 sed -i 's/ /\t/g' $DIR/input/*
 
-if [ ! -e $DIR/visual_id_file.tsv ]; then
-	gdown '1mUZ34Jx9H5LqnguGZuX2-0G_PBi4kAiS' -O $DIR/visual_id_file.tsv
+if [[ $TSV == 'skip' ]]; then
+    :
+else
+    rm -f $DIR/prime-output/*
+    if [ ! -e $TSV ]; then
+        echo "TSV directory not found: $TSV"
+        exit 1
+    fi
+
+    if [[ $TSV == *"v3"* ]]; then
+        python $DIR/de_duplicate.py -qre $QREL -tsv $TSV -sub "$DIR/input/" -v 6 -pri "$DIR/prime-output/"
+    elif [[ $TSV == *"v2"* ]]; then
+        python $DIR/de_duplicate.py -qre $QREL -tsv $TSV -sub "$DIR/input/" -v 4 -pri "$DIR/prime-output/"
+    else
+        echo "Cannot guess the version of TSV directory: $TSV"
+        exit 1
+    fi
 fi
-if [ ! -e $DIR/formulas_slt_string.tsv ]; then
-	gdown '1nRrG3T2hrQY-VU0awoHSG7-g8W-6pXth' -O $DIR/formulas_slt_string.tsv
-fi
 
-python $DIR/arqmath_2020_task2_convert_runs.py -ru "$DIR/input/" -re "$DIR/prime-output/" -v $DIR/visual_id_file.tsv -q $QREL -ld $TSV -s $DIR/formulas_slt_string.tsv
-
-python $DIR/task2_get_results.py -eva trec_eval -qre $QREL -de "$DIR/prime-output/" -res $DIR/result.tsv $NOJUDGE
-
+python $DIR/task2_get_results.py -eva trec_eval -qre $QREL -pri "$DIR/prime-output/" -res $DIR/result.tsv $NOJUDGE
 cat $DIR/result.tsv | sed -e 's/[[:blank:]]/ /g'
